requirements.txt
torch>=2.0.0
transformers>=4.38.0
peft>=0.9.0
bitsandbytes>=0.41.0
sentence-transformers>=2.5.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
nltk>=3.8.1
faiss-gpu>=1.7.4
tqdm>=4.66.0
accelerate>=0.27.2
